Vinisha, Rakesh | MOBN-1376 | Adds tests and refactoring
